fix(aurora): refactored the code to avoid duplication
